flask/requirements/prod.txt
flask
pandas
sklearn
python-dotenv
flask-restx
uwsgi
redis
pytest